requirements.txt
Flask==2.2.2
pycryptodomex==3.11.0
Werkzeug==2.2.2